+++ astropy_xarray/coordinates/sky_coord.py
@@ -5,9 +5,10 @@ import astropy.units as u
 import numpy as np
 import xarray as xr
 from astropy.coordinates import SkyCoord
+from astropy.time import Time
 from astropy.utils import ShapedLikeNDArray
 
-from astropy_xarray.coordinates.core import dump_time
+from astropy_xarray.coordinates.core import dump_time, load_optional_object
 from astropy_xarray.coordinates.frame import dump_frame, load_frame, load_representation
 
 _ArrayLike = list | np.ndarray | ShapedLikeNDArray
@@ -110,6 +111,11 @@ def skycoord_to_dataset(
             dict(obstime=dump_time(skycoord.obstime))
             if skycoord.obstime is not None and not hasattr(skycoord.frame, "obstime")
             else {}
+        )
+        | (
+            dict(equinox=dump_time(skycoord.equinox))
+            if skycoord.equinox is not None and not hasattr(skycoord.frame, "equinox")
+            else {}
         ),
     )
 
@@ -133,4 +139,14 @@ def dataset_to_skycoord(ds: xr.Dataset) -> SkyCoord:
     )
     frame.representation_type = ds.attrs["frame"]["representation_type"]
     frame.differential_type = ds.attrs["frame"].get("differential_type")
-    return SkyCoord(frame)
+
+    kwargs = {
+        attr: value
+        for attr in ("obstime", "equinox")
+        if not hasattr(frame, attr)
+        and (value := load_optional_object(Time, ds.attrs.get(attr))) is not None
+    }
+    return SkyCoord(
+        frame,
+        **kwargs,
+    )